requirements.txt
python-chess
torch 
torchvision 
stable-baselines3 
gym 